HW08. Fix of the package "model", fix of the ResumeTest main-class, fix of the AbstractFileStorage class

+++ src/ru/javawebinar/basejava/storage/AbstractFileStorage.java
@@ -37,9 +37,12 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
 
     @Override
     public void clear() {
-        for (File file : directory.listFiles()) {
-            if (file.isFile()) {
-                file.delete();
+        File[] files = directory.listFiles();
+        if (files != null) {
+            for (File file : files) {
+                if (file.isFile()) {
+                    file.delete();
+                }
             }
         }
     }
@@ -61,8 +64,11 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
     @Override
     protected List<Resume> getAllStorage() {
         List<Resume> resumes = new ArrayList<>();
-        for (File file : directory.listFiles()) {
-            resumes.add(getFromStorage(file));
+        File[] files = directory.listFiles();
+        if (files != null) {
+            for (File file : files) {
+                resumes.add(getFromStorage(file));
+            }
         }
         return resumes;
     }
@@ -80,7 +86,12 @@ public abstract class AbstractFileStorage extends AbstractStorage<File> {
 
     @Override
     public int size() {
-        return directory.listFiles().length;
+        File[] files = directory.listFiles();
+        int length = files.length;
+        if (files != null) {
+            return length;
+        }
+        return 0;
     }
 
     @Override
